refactor(mobile): remove shared module
