+++ mobile/lib/utils/hooks/app_settings_update_hook.dart
